refactor: fix encoding and move str2bool to shared utils
- Add encoding="utf-8" to open() in tn/main.py and itn/main.py for
  Windows compatibility
- Move str2bool from itn/main.py to tn/utils.py to remove TN's
  dependency on ITN module
